+++ db/users.js
@@ -6,7 +6,7 @@ module.exports = {
   authenticate,
   getByName,
   getByID,
-  getByEmail,
+//   getByEmail,
   setPassword,
   setEmail,
   purgeByName,
@@ -14,15 +14,20 @@ module.exports = {
 
 // returns user
 async function create(username, password, email) {
-  if (!username) {
-    throw new Error("Create user requires username");
+  if (!username || typeof username !== 'string') {
+    throw new TypeError("Create user requires username");
   }
+  username = username.trim();
   if (await getByName(username)) { // might be inefficient, but
     throw new Error("Username already exists");
   }
   email = email || null; // explicit sql null
+  if (email) {
+    email = email.trim();
+  }
   let hashedPassword = null;
-  if (password) {
+  if (password && typeof password === 'string') {
+    password = password.trim();
     hashedPassword = await bcrypt.hash(password, 12);
   }
   const result = await db.query("INSERT INTO users (username, password, email) VALUES ($1, $2, $3) RETURNING id, username, email, created", [username, hashedPassword, email]);
@@ -30,9 +35,11 @@ async function create(username, password, email) {
 }
 
 async function authenticate(username, password) {
-  if (!username || !password) {
-    throw new Error("Authenticate user requires username and password");
+  if (!username || !password || typeof username !== 'string' || typeof password !== 'string') {
+    throw new TypeError("Authenticate user requires username and password");
   }
+  username = username.trim();
+  password = password.trim();
   const result = await db.query("SELECT password FROM users WHERE username = $1", [username]);
   try {
     const hashedPassword = result.rows[0].password;
@@ -47,26 +54,37 @@ async function authenticate(username, password) {
 }
 
 async function getByName(username) {
-  const result = await db.query("SELECT id, username, email, created FROM users WHERE username = $1", [username]);
-  return result.rows[0];
+  if (typeof username === 'string') {
+    username = username.trim();
+    const result = await db.query("SELECT id, username, email, created FROM users WHERE username = $1", [username]);
+    return result.rows[0];
+  } else {
+    throw new TypeError("Get user by name requires username");
+  }
 }
 
 async function getByID(id) {
-  return db.query("SELECT id, username, email, created FROM users WHERE id = $1", [id]).then(result => result.rows[0]);
+  if (typeof id === 'number') {
+    return db.query("SELECT id, username, email, created FROM users WHERE id = $1", [id]).then(result => result.rows[0]);
+  } else {
+    throw new TypeError("Get user by ID requires numeric ID");
+  }
 }
 
 // returns (possibly empty) array, probably won't use this except for manual admin
-async function getByEmail(email) {
-  return db.query("SELECT id, username, email, created FROM users WHERE email = $1", [email]).then(result => result.rows);
-}
+// async function getByEmail(email) {
+//   return db.query("SELECT id, username, email, created FROM users WHERE email = $1", [email]).then(result => result.rows);
+// }
 
 // application logic is in charge of validating, this is raw
 async function setPassword(username, password) {
-  if (!username) {
-    throw new Error("setPassword requires username"); // but blanking pw is ok
+  if (!username || typeof username !== 'string') {
+    throw new TypeError("setPassword requires username"); // but blanking pw is ok
   }
+  username = username.trim();
   let hashedPassword = null; // explicit sql null
-  if (password) {
+  if (password && typeof password === 'string') {
+    password = password.trim();
     hashedPassword = await bcrypt.hash(password, 12);
   }
   await db.query("UPDATE users SET password = $2 WHERE username = $1", [username, hashedPassword]);
@@ -74,12 +92,23 @@ async function setPassword(username, password) {
 
 // returns user
 async function setEmail(username, email) {
+  if (!username || typeof username !== 'string') {
+    throw new TypeError("setEmail requires username"); // but blanking pw is ok
+  }
+  username = username.trim();
   email = email || null; // explicit sql null
+  if (email && typeof email === 'string') {
+    email = email.trim();
+  }
   return db.query("UPDATE users SET email = $2 WHERE username = $1 RETURNING id, username, email, created", [username, email]).then(result => result.rows[0]);
 }
 
 // schema should just make this cascade to dogears.
 async function purgeByName(username) {
+  if (!username || typeof username !== 'string') {
+    throw new TypeError("purgeByName requires username"); // but blanking pw is ok
+  }
+  username = username.trim();
   await db.query("DELETE FROM users WHERE username = $1", [username]);
 }
